requirements.txt
fastapi
uvicorn
pydantic
requests
ibm-cloud-sdk-core
python-dotenv
pydantic[email]
python-multipart
jinja2